Work I did over a couple of hours
Uses a 'bad' detected set of points to compare to original, and then trigger to send to the Arudino which will see the RX light on for incoming data, and ALSO the data detected corresponds to a value which will trigger the onboard LED. The thinking is in the future, you can use the values to trigger multiple separate things I guess. Either way, its processed data sent back to do something. Will need to look into how this could be done in real-time with minimal delay. ie Signal comes in, onsets are detected (with decent accuracy) and then things happen, vs currently, Signal goes to data-set, data-set is analyzed and then things sent out. I also added a .5sec slow just so its easier to see positive signal separation on detection.